sound test direct Sound in C
git-svn-id: svn+ssh://blue.3dallee.de/fhtw/trunk@14 6aeb069d-d5e0-465d-890a-ccf59a9091f3

+++ AirGapC/ClockRecovery.h
@@ -1,2 +1,2 @@
 void(*ClockRecovery_ReportData)(short *);
-void ClockRecovery_OnData(short *data);
+void ClockRecovery_OnData(float *data);
